feat(util): 添加显示上下文菜单工具类，全局使用单例模式
